% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chill_func}
\alias{chill_func}
\title{Chill portion calculator}
\usage{
chill_func(temp)
}
\arguments{
\item{temp}{a vector of temperautres \emph{in Celsius}}
}
\value{
The output is the chill portion calculation. Portions are rounded
down to the largest integer.
}
\description{
\code{chill_func} returns the chill portion count of a temperature vector
}
